requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
django-filter==23.3
celery==5.3.1
redis==4.6.0
psycopg2-binary==2.9.7
python-decouple==3.8
drf-yasg==1.21.7
django-auditlog==2.3.1
Pillow==10.0.0
whitenoise==6.5.0
gunicorn==21.2.0
djangorestframework-simplejwt==5.3.0
django-extensions==3.2.3